+++ CHANGES.md
@@ -2,10 +2,11 @@
 
 ## 0.11.0
 
+- feat(`getSectionInfoForWork`): add new method
+- feat(`getParagraphsForSectionId`): add new method
 - fix(`getSectionNamesForWork`): ensure returns from `mainSections` if not
   found in `subSections` (will get first lines)
 - fix(`getUrlForWorkAndSection`): gets the paragraphs for a work and first line
-- feat(`getSectionInfoForWork`): add new method
 
 ## 0.10.0